Split sport config form into modular folder
